+++ isitfit/cli.py
@@ -61,6 +61,7 @@ def cli(debug, version, optimize, n, filter_tags):
         mm.add_listener('all', ul.after_all)
         mm.add_listener('all', ul.display_all)
       else:
+        mm.add_listener('pre', ol.handle_pre)
         mm.add_listener('ec2', ol.per_ec2)
         mm.add_listener('all', ol.after_all)
         mm.add_listener('all', ol.display_all)